backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy[asyncio]
asyncpg
aiosqlite
alembic
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
stripe
celery
redis
httpx
numpy
pandas
scikit-learn
xgboost
pulp
slowapi
python-multipart
email-validator
pytest
pytest-asyncio
ruff
